move base path computation out of closure
